:recycle: 결제 관련 리팩토링

+++ src/main/java/kr/hhplus/concert/application/dto/ReservationCommand.java
@@ -6,7 +6,7 @@ import lombok.Builder;
 public record ReservationCommand (
     Long userId,
     Long concertId,
-    Integer concertScheduleId,
-    Integer seatId
+    Long concertScheduleId,
+    Long seatId
 ){
 }
